Fix variable

+++ migration/migrator/migrations/master/20240718015237_self_account_creation.py
@@ -1,7 +1,7 @@
 """Migration for the Submitty master database."""
 
 import json
-
+from pathlib import Path
 def up(config, database):
     """
     Run up migration.
@@ -11,13 +11,16 @@ def up(config, database):
     :param database: Object for interacting with given database for environment
     :type database: migrator.db.Database
     """
-    with open('/usr/local/submitty/config/submitty.json', 'r') as conf:
-        SUBMITTY_CONFIG_JSON = json.load(conf)
-    if 'user_create_account' not in SUBMITTY_CONFIG_JSON:
-        SUBMITTY_CONFIG_JSON['user_create_account'] = False
-    
-    dump = open('/usr/local/submitty/config/submitty.json', 'w')
-    json.dump(SUBMITTY_CONFIG_JSON, dump, indent=4)
+
+    my_file = Path("/usr/local/submitty/config/submitty.json")
+    if my_file.is_file():
+        with open('/usr/local/submitty/config/submitty.json', 'r') as conf:
+            SUBMITTY_CONFIG_JSON = json.load(conf)
+        if 'user_create_account' not in SUBMITTY_CONFIG_JSON:
+            SUBMITTY_CONFIG_JSON['user_create_account'] = False
+        
+        dump = open('/usr/local/submitty/config/submitty.json', 'w')
+        json.dump(SUBMITTY_CONFIG_JSON, dump, indent=4)
 
 
 def down(config, database):
